Commit Con merketplace e carte

+++ src/app/componenti/pagina-marketplace/pagina-marketplace.component.ts
@@ -23,6 +23,7 @@ export class PaginaMarketplaceComponent implements OnInit{
   anno_pubblicazione:string=""
   itemsA: any
   itemsT:any
+  costo:any;
   private baseUrl = 'http://localhost:9092/api/v1/cartemazzi'
   private baseurl2 ='http://localhost:9091/api/v1/marketplace'
   // @ts-ignore
@@ -36,12 +37,15 @@ export class PaginaMarketplaceComponent implements OnInit{
   selectedAnnoPubblicazione: any;
   selectedCosto: any;
   id:any;
+  private nickname:any;
+  private risposta:any;
   constructor(private show: ShowCarteInVenditaService,private dialog: MatDialog, private buy:AcquistaService,private http:HttpClient,private nick:Nickname_and_email_user_loggedService) {
 
   }
 
   ngOnInit(): void {
     this.fetchCarteInVendita();
+    console.log(this.nick.getStoredNickname_user_logged())
   }
 
   fetchCarteInVendita(){
@@ -66,6 +70,7 @@ export class PaginaMarketplaceComponent implements OnInit{
     const dialogRef = this.dialog.open(PopupComponent, {
       data: carta,
     });
+    console.log(carta.nick)
     dialogRef.afterClosed().subscribe(result => {
       if (result === 'acquisto-confermato') {
         console.log("confermato");
@@ -73,22 +78,26 @@ export class PaginaMarketplaceComponent implements OnInit{
         let url: string;
         let url2: string;
         if (carta.hasOwnProperty('durata')) {
-          url = `${this.baseUrl}/acquistaCartaBrano`;
+          url = `${this.baseUrl}/acquistaCartaBrano/${this.nick.getStoredNickname_user_logged()}`;
           url2 = `${this.baseurl2}/delete-CardBrano/${carta.id}`
           this.id=carta.id;
           this.nome = carta.nome;
           this.durata = carta.durata;
           this.anno_pubblicazione = carta.anno_pubblicazione;
           this.popolarita = carta.popolarita;
+          this.costo = carta.costo;
           this.immagine = carta.immagine;
+          this.nickname=carta.nick;
         } else {
-          url = `${this.baseUrl}/acquistaCartaArtista`;
+          url = `${this.baseUrl}/acquistaCartaArtista/${this.nick.getStoredNickname_user_logged()}`;
           url2 = `${this.baseurl2}/delete-CardArtist/${carta.id}`
           this.id=carta.id;
           this.nome = carta.nome;
           this.popolarita = carta.popolarita;
           this.genere = carta.genere;
+          this.costo = carta.costo;
           this.immagine = carta.immagine;
+          this.nickname=carta.nick;
         }
         if (carta.hasOwnProperty('durata')) {
           this.http.post(url, {
@@ -98,62 +107,75 @@ export class PaginaMarketplaceComponent implements OnInit{
             "anno_pubblicazione": this.anno_pubblicazione,
             "popolarita": this.popolarita,
             "immagine": this.immagine,
-            "nickname":this.nick.getStoredNickname_user_logged()
+            "costo": this.costo,
+            "nick":this.nickname
           })
             .subscribe(
               response => {
                 // Gestisci la risposta dal backend
                 console.log("Risposta dal backend:", response);
+
               },
               error => {
                 // Gestisci gli errori
                 console.error("Errore nella chiamata:", error);
+
               }
             );
-          this.http.delete(url2)
-      .subscribe(
-          response => {
-            // Gestisci la risposta dal backend
-            console.log("Risposta dal backend:", response);
-          },
-          error => {
-            // Gestisci gli errori
-            console.error("Errore nella chiamata:", error);
+          if(this.risposta != "Punti insufficienti per acquistare questa carta.") {
+            this.http.delete(url2)
+              .subscribe(
+                response => {
+                  // Gestisci la risposta dal backend
+                  console.log("Risposta dal backend:", response);
+                },
+                error => {
+                  // Gestisci gli errori
+                  console.error("Errore nella chiamata:", error);
+                }
+              );
+            this.removeItemFromArray(carta);
+
           }
-        );
         } else {
           this.http.post(url, {
-            "id":this.id,
+            "id": this.id,
             "nome": this.nome,
             "popolarita": this.popolarita,
             "genere": this.genere,
             "immagine": this.immagine,
-            "nickname":this.nick.getStoredNickname_user_logged()
+            "costo": this.costo,
+            "nick": this.nickname
           })
             .subscribe(
               response => {
                 // Gestisci la risposta dal backend
                 console.log("Risposta dal backend:", response);
+
               },
               error => {
                 // Gestisci gli errori
                 console.error("Errore nella chiamata:", error);
+
               }
             );
-          console.log(url2)
-          this.http.delete(url2)
-            .subscribe(
-              response => {
-                // Gestisci la risposta dal backend
-                console.log("Risposta dal backend:", response);
-              },
-              error => {
-                // Gestisci gli errori
-                console.error("Errore nella chiamata:", error);
-              }
-            );
+          console.log("OH sto qua:"+this.risposta)
+          if (this.risposta != "Punti insufficienti per acquistare questa carta.") {
+            this.http.delete(url2)
+              .subscribe(
+                response => {
+                  // Gestisci la risposta dal backend
+                  console.log("Risposta dal backend:", response);
+                },
+                error => {
+                  // Gestisci gli errori
+                  console.error("Errore nella chiamata:", error);
+                }
+              );
+            this.removeItemFromArray(carta);
+          }
+
         }
-        this.removeItemFromArray(carta);
 
 
       }
